went back to loading php pages, didnt commit js handling attempts

+++ part1/delete.php
@@ -3,7 +3,13 @@
 include("connection.php");
 
 $id = $_GET['id']; 
+echo $id;
 
-$query = mysqli_query($connection,"delete from bookmarks where bookmark_id = '$id'");
+if($_SERVER['REQUEST_METHOD'] == "GET" && isset($_GET["id"])){
+    $query = mysqli_query($connection,"delete from bookmarks where bookmark_id = '$id'");
+    echo "HERE";
+}
 
+header("Location: home.php");
+die;
 ?>
